+++ worker-email/src/types.ts
